LangChainChatModel done
