+++ experimental/pom.xml
@@ -35,6 +35,11 @@
             <artifactId>serverlessworkflow-fluent-func</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>io.serverlessworkflow</groupId>
+            <artifactId>serverlessworkflow-experimental-ai-types</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
   </dependencyManagement>
   <modules>
